+++ models/intermediate/int_shopify__daily_fulfillment.sql
@@ -12,7 +12,7 @@ fulfillment_aggregates as (
         source_relation,
         cast({{ dbt.date_trunc('day','happened_at') }} as date) as date_day
 
-        {% for status in ['attempted_delivery', 'delivered', 'failure', 'in_transit', 'out_for_delivery', 'ready_for_pickup', 'picked_up', 'label_printed', 'label_purchased', 'confirmed']%}
+        {% for status in ['attempted_delivery', 'delayed', 'delivered', 'failure', 'in_transit', 'out_for_delivery', 'ready_for_pickup', 'picked_up', 'label_printed', 'label_purchased', 'confirmed']%}
         , count(distinct case when lower(status) = '{{ status }}' then fulfillment_id end) as count_fulfillment_{{ status }}
         {% endfor %}